Fixed windows build regressions.

+++ examples/glEvalLimit/glEvalLimit.cpp
@@ -427,7 +427,7 @@ createOsdMesh(ShapeDesc const & shapeDesc, int level) {
         if (g_drawMode==kVARYING) {
             g_varyingData = Osd::CpuVertexBuffer::Create(3, nverts);
             g_varyingData->UpdateData( 
-                &g_varyingColors[0], 0, g_varyingColors.size()/3 );
+                &g_varyingColors[0], 0, (int)g_varyingColors.size()/3 );
         }
 
         // Create output buffers for the limit samples (position & tangents)
